backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
boto3==1.28.64
python-dotenv==1.0.0
httpx==0.23.3
supabase==1.0.3
pytest==7.4.3
pydantic==1.10.13
python-jose==3.3.0
flake8==6.1.0
starlette==0.27.0
typing-extensions==4.8.0
anyio==3.7.1
pandas>=2.1.0
openpyxl==3.1.2